+++ plugins/azure/src/components/costmanagement/helpers.ts
@@ -1,0 +1,13 @@
+import { IOptions } from './interfaces';
+import { getTimeParams } from '@kobsio/plugin-core';
+
+// getInitialOptions returns the initial options for the applications page from the url.
+export const getInitialOptions = (search: string, isInitial: boolean): IOptions => {
+  const params = new URLSearchParams(search);
+  const scope = params.get('scope');
+
+  return {
+    scope: scope ? scope : 'All',
+    times: getTimeParams(params, isInitial),
+  };
+};